fetching author
name timestamp and avatar updates on feed.
