+++ js/index.mjs
@@ -1,6 +1,6 @@
 import * as templates from "./templates/index.mjs";
 import * as postsMethods from "./api/posts/index.mjs";
-
+import * as utils from "./utils/index.mjs";
 import * as listeners from "./handlers/index.mjs";
 
 document.addEventListener("DOMContentLoaded", () => {
@@ -9,6 +9,7 @@ document.addEventListener("DOMContentLoaded", () => {
 
   if (path === "/profile/login/index.html") {
     listeners.setLoginFormListener();
+
   } else if (path === "/profile/register/") {
     listeners.setRegisterFormListener();
   } else if (path === "/profile/edit/") {
@@ -17,6 +18,10 @@ document.addEventListener("DOMContentLoaded", () => {
     listeners.setUpdatePostListener();
   } else if (path === "/feed/index.html") {
     listeners.displayPosts();
-    listeners.setCreatePostListener(); 
+    listeners.setCreatePostListener();
   }
+
+  // utils.profileUpdater.init();
+  // feedUpdater.init();
+  utils.updateProfileFromLocalStorage();
 });
