+++ js/utils/index.mjs
@@ -1,2 +1,3 @@
 export * from "./messages.mjs"
 export * from "./profileUpdater.mjs"
+export * from "./feedUpdater.mjs"
