+++ js/utils/profileUpdater.mjs
@@ -1,6 +1,6 @@
 import { load } from "../storage/index.mjs";
 
-document.addEventListener("DOMContentLoaded", function () {
+export function updateProfileFromLocalStorage() {
   // Retrieve user profile from localStorage
   const userProfile = load("profile");
 
@@ -17,9 +17,12 @@ document.addEventListener("DOMContentLoaded", function () {
     // Replace "img" with the user's avatar image URL or default avatar if not provided
     const elementsWithAvatar = document.querySelectorAll(".replace-avatar");
     elementsWithAvatar.forEach((element) => {
-      const avatarImage = userProfile.avatar ? userProfile.avatar : "../../image/default-avatar.JPG";
+      const avatarImage = userProfile.avatar
+        ? userProfile.avatar
+        : "../../image/default-avatar.JPG";
       element.src = avatarImage;
-      element.alt = "profile picture default - silhouette of a person in profile with paint around head";
+      element.alt =
+        "profile picture default - silhouette of a person in profile with paint around head";
     });
 
     // Set banner image
@@ -33,4 +36,6 @@ document.addEventListener("DOMContentLoaded", function () {
       }
     }
   }
-});
+}
+// Execute profile update logic when DOM content is loaded
+document.addEventListener("DOMContentLoaded", updateProfileFromLocalStorage);
