+++ profile/index.html
@@ -9,7 +9,7 @@
   <script src="https://kit.fontawesome.com/db895f6c4c.js" crossorigin="anonymous"></script>
   <link rel="stylesheet" href="../css/styles.css">
   <script src="../js/hamburger.js"></script>
-  <script src="../js/utils/profileUpdater.mjs" defer type="module"></script>
+  <script src="../js/index.mjs" defer type="module"></script>
   <title>Profile</title>
 </head>
